fix(integer-i64): widen stdlib #rust annotations from i32 to i64
Unblocks runtime — minimal println("hello") now works, probes 08 + 12
pass, probe 11 has 9/12 passing (remaining 3 are expected test-tail
failures from i32::MAX overflow tests that no longer apply at i64).

Stdlib annotations (default/01_code.loft):
- OpConstShort / OpConstTiny: i32::from → i64::from.
- OpVarInt: get_var::<i32> → get_var::<i64>.
- OpConvIntFromNull, OpCastIntFromText, OpGetInt, OpCastEnumFromInt:
  i32::MIN → i64::MIN.
- OpConvIntFromCharacter: i32::MIN → i64::MIN, `@v1 as i32` →
  `i64::from(@v1 as u32)`.
- OpConvIntFromEnum: i32::MIN / i32::from → i64::MIN / i64::from.
- OpLengthText / OpSizeText / OpLengthVector: `as i32` → `as i64`.
- OpSetRef / OpSetCharacter / OpSetText / OpGetCharacter / OpGetText:
  role-2 writes/reads of u32 slots → set_u32_raw / get_u32_raw.
- OpGetByte / OpGetShort: wrap return in `i64::from(...)`.
- OpSetByte / OpSetShort: cast `@val` (now i64) back to i32 for the
  store-side clamp.

Stdlib annotations (default/03_text.loft):
- find / rfind: `v as i32` → `v as i64`, `i32::MIN` → `i64::MIN`.

Runtime handlers (hand-maintained alongside fill.rs):
- src/native.rs n_assert, n_panic, n_log_info/warn/error/fatal:
  `v_line = *stores.get::<i32>(...)` → `get::<i64>` — matches `line:
  integer` in stdlib signatures.
- src/state/text.rs format_text / format_stack_text:
  `get_stack::<i32>` (width) → `get_stack::<i64> as i32`; string_mut
  pop offset: `pos - 4 - size_ptr()` → `pos - 8 - size_ptr()`.
- src/state/text.rs format_long / format_float / format_single and
  their _stack variants: same widening; string pop offsets
  recomputed for 8-byte integer args.
- src/state/text.rs get_text_sub: `get_stack::<i32>` (till, from) →
  `get_stack::<i64> as i32`.
- src/state/mod.rs fn_call_ref: fn-ref slot widens from 16 → 20
  bytes to accommodate i64 d_nr ([d_nr:i64][closure:DbRef]).
  Reader uses `get_var::<i64>`; closure offset `fn_var - 4` →
  `fn_var - 8`.
- src/state/debug.rs dump_attribute / log path: `get_stack::<i32>`
  → `get_stack::<i64>` for `Type::Integer`.
- src/state/io.rs: `read_file` bytes param, key builders for
  Content::Long (tp=0|1|6), append_copy multiply, insert_vector
  index — all widened to i64 at the stack read, narrowed back at
  downstream calls that still take i32/u32.
- src/state/io.rs: assorted `get_int(...) as u32` for vector
  headers / hash buckets — replaced with `get_u32_raw` (role-2).

Regenerated src/fill.rs.

+++ src/native.rs
@@ -134,7 +134,7 @@ pub fn init(state: &mut State) {
 }
 
 fn n_assert(stores: &mut Stores, stack: &mut DbRef) {
-    let v_line = *stores.get::<i32>(stack);
+    let v_line = *stores.get::<i64>(stack);
     let v_file = *stores.get::<Str>(stack);
     let v_message = *stores.get::<Str>(stack);
     let v_test = *stores.get::<bool>(stack);
@@ -174,7 +174,7 @@ fn n_assert(stores: &mut Stores, stack: &mut DbRef) {
 }
 
 fn n_panic(stores: &mut Stores, stack: &mut DbRef) {
-    let v_line = *stores.get::<i32>(stack);
+    let v_line = *stores.get::<i64>(stack);
     let v_file = *stores.get::<Str>(stack);
     let v_message = *stores.get::<Str>(stack);
     if let Some(ref logger) = stores.logger {
@@ -198,7 +198,7 @@ fn n_panic(stores: &mut Stores, stack: &mut DbRef) {
 }
 
 fn n_log_info(stores: &mut Stores, stack: &mut DbRef) {
-    let v_line = *stores.get::<i32>(stack);
+    let v_line = *stores.get::<i64>(stack);
     let v_file = *stores.get::<Str>(stack);
     let v_message = *stores.get::<Str>(stack);
     if let Some(ref logger) = stores.logger
@@ -209,7 +209,7 @@ fn n_log_info(stores: &mut Stores, stack: &mut DbRef) {
 }
 
 fn n_log_warn(stores: &mut Stores, stack: &mut DbRef) {
-    let v_line = *stores.get::<i32>(stack);
+    let v_line = *stores.get::<i64>(stack);
     let v_file = *stores.get::<Str>(stack);
     let v_message = *stores.get::<Str>(stack);
     if let Some(ref logger) = stores.logger
@@ -220,7 +220,7 @@ fn n_log_warn(stores: &mut Stores, stack: &mut DbRef) {
 }
 
 fn n_log_error(stores: &mut Stores, stack: &mut DbRef) {
-    let v_line = *stores.get::<i32>(stack);
+    let v_line = *stores.get::<i64>(stack);
     let v_file = *stores.get::<Str>(stack);
     let v_message = *stores.get::<Str>(stack);
     if let Some(ref logger) = stores.logger
@@ -236,7 +236,7 @@ fn n_log_error(stores: &mut Stores, stack: &mut DbRef) {
 }
 
 fn n_log_fatal(stores: &mut Stores, stack: &mut DbRef) {
-    let v_line = *stores.get::<i32>(stack);
+    let v_line = *stores.get::<i64>(stack);
     let v_file = *stores.get::<Str>(stack);
     let v_message = *stores.get::<Str>(stack);
     if let Some(ref logger) = stores.logger
